requirements.txt
requests
beautifulsoup4
streamlit
networkx
plotly
pandas
lxml
openai
urllib3
sqlalchemy
python-dotenv
validators
tqdm
numpy
selenium
webdriver-manager   